Extract AbstractYadasModel (story #491)
