+++ mtc-mcmc/src/integration-test/java/org/drugis/mtc/yadas/ContinuousDataIT.java
@@ -29,7 +29,7 @@ import javax.xml.bind.JAXBException;
 import org.apache.commons.math3.stat.descriptive.moment.Mean;
 import org.apache.commons.math3.stat.descriptive.moment.StandardDeviation;
 import org.drugis.common.threading.TaskUtil;
-import org.drugis.mtc.MixedTreatmentComparison.ExtendSimulation;
+import org.drugis.mtc.MCMCModel.ExtendSimulation;
 import org.drugis.mtc.ResultsUtil;
 import org.drugis.mtc.model.JAXBHandler;
 import org.drugis.mtc.model.Network;
