+++ mtc-mcmc/src/integration-test/java/org/drugis/mtc/yadas/YadasInconsistencyModelIT.java
@@ -21,7 +21,7 @@ package org.drugis.mtc.yadas;
 
 import org.drugis.mtc.InconsistencyModel;
 import org.drugis.mtc.InconsistencyModelTestBase;
-import org.drugis.mtc.MixedTreatmentComparison.ExtendSimulation;
+import org.drugis.mtc.MCMCModel.ExtendSimulation;
 import org.drugis.mtc.model.Network;
 
 public class YadasInconsistencyModelIT extends InconsistencyModelTestBase {
