+++ mtc-mcmc/src/main/java/org/drugis/mtc/MixedTreatmentComparison.java
@@ -27,16 +27,6 @@ import org.drugis.mtc.model.Treatment;
  * the model has been run (by executing the ActivityModel).
  */
 public interface MixedTreatmentComparison extends MCMCModel {
-	public static final String ASSESS_CONVERGENCE_PHASE = "Assess convergence";
-	public static final String CALCULATING_SUMMARIES_PHASE = "Calculating summaries";
-	public static final String EXTENDING_SIMULATION_PHASE = "Extending simulation";
-	public static final String STARTING_SIMULATION_PHASE = "Building model";
-
-	
-	public enum ExtendSimulation { 
-		WAIT, EXTEND, FINISH
-	}
-	
 	/**
 	 * Get the relative effect MCMCParameter.
 	 * @return The effect parameter.
@@ -52,12 +42,5 @@ public interface MixedTreatmentComparison extends MCMCModel {
 	 * Get the random effects variance.
 	 * @return The variance parameter.
 	 */
-	public Parameter getRandomEffectsVariance();
-
-	
-	/**
-	 * @param s Whether to finish or extend the simulation, or to wait for input.
-	 */
-	public void setExtendSimulation(ExtendSimulation s);
-	
+	public Parameter getRandomEffectsVariance();	
 }
