+++ mtc-mcmc/src/main/java/org/drugis/mtc/yadas/YadasModel.java
@@ -28,9 +28,7 @@ import gov.lanl.yadas.Gaussian;
 import gov.lanl.yadas.GroupArgument;
 import gov.lanl.yadas.IdentityArgument;
 import gov.lanl.yadas.MCMCParameter;
-import gov.lanl.yadas.MCMCUpdate;
 import gov.lanl.yadas.Uniform;
-import gov.lanl.yadas.UpdateTuner;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -39,24 +37,6 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.commons.math3.random.JDKRandomGenerator;
-import org.drugis.common.threading.AbstractExtendableIterativeComputation;
-import org.drugis.common.threading.AbstractIterativeComputation;
-import org.drugis.common.threading.ExtendableIterativeTask;
-import org.drugis.common.threading.IterativeTask;
-import org.drugis.common.threading.NullTask;
-import org.drugis.common.threading.SimpleSuspendableTask;
-import org.drugis.common.threading.Task;
-import org.drugis.common.threading.WaitingTask;
-import org.drugis.common.threading.activity.ActivityModel;
-import org.drugis.common.threading.activity.ActivityTask;
-import org.drugis.common.threading.activity.Condition;
-import org.drugis.common.threading.activity.DecisionTransition;
-import org.drugis.common.threading.activity.DirectTransition;
-import org.drugis.common.threading.activity.ForkTransition;
-import org.drugis.common.threading.activity.JoinTransition;
-import org.drugis.common.threading.activity.Transition;
-import org.drugis.mtc.MCMCResults;
-import org.drugis.mtc.MCMCSettingsCache;
 import org.drugis.mtc.MixedTreatmentComparison;
 import org.drugis.mtc.Parameter;
 import org.drugis.mtc.model.Measurement;
@@ -77,229 +57,31 @@ import org.drugis.mtc.parameterization.RandomEffectsVariance;
 import org.drugis.mtc.parameterization.SplitParameter;
 import org.drugis.mtc.parameterization.StartingValueGenerator;
 
-abstract class YadasModel implements MixedTreatmentComparison {
-	private static final int THINNING_INTERVAL = 1;
-	private static final double VARIANCE_SCALING = 2.5;
+abstract class YadasModel extends AbstractYadasModel implements MixedTreatmentComparison {
 	protected final Network d_network;
 	protected Parameterization d_pmtz = null;
 	
-	private PriorGenerator d_priorGen;
+	PriorGenerator d_priorGen;
 	protected List<StartingValueGenerator> d_startGen = new ArrayList<StartingValueGenerator>();
-	protected final int d_nChains = 4;
-	
-	private List<List<ParameterWriter>> d_writeList = new ArrayList<List<ParameterWriter>>();
-	private List<List<MCMCUpdate>> d_updateList = new ArrayList<List<MCMCUpdate>>();
-
 	protected Parameter d_randomEffectVar = new RandomEffectsVariance();
 	protected Parameter d_inconsistencyVar = new InconsistencyVariance();
-
-	private int d_burnInIter = 20000;
-	protected int d_simulationIter = 60000;
-	private int d_reportingInterval = 100;
-
-	private YadasResults d_results = new YadasResults();
-	private ActivityTask d_activityTask;
-
-	private SimpleSuspendableTask d_finalPhase;
-	private ExtendSimulation d_extendSimulation = ExtendSimulation.WAIT;
-	private Task d_extendDecisionPhase;
-	private Task d_extendSimulationPhase;
-	private SimpleSuspendableTask d_notifyResults;
-	
-	private final class ExtendDecisionTask extends WaitingTask {
-		@Override
-		public boolean isWaiting() {
-			return d_extendSimulation == ExtendSimulation.WAIT;
-		}
-
-		@Override
-		public void onEndWaiting() {
-			d_started = false;
-			d_finished = true;
-			if (d_extendSimulation == ExtendSimulation.EXTEND) {
-				((SimpleRestartableSuspendableTask) d_extendSimulationPhase).reset();
-			}
-			d_mgr.fireTaskFinished();
-		}
-		
-		public void reset() {
-			d_extendSimulation = ExtendSimulation.WAIT;
-			d_finished = false;
-			d_mgr.fireTaskRestarted();
-		}
-		
-		@Override
-		public String toString() {
-			return MixedTreatmentComparison.ASSESS_CONVERGENCE_PHASE;
-		}
-	}
-
-	private class BurnInChain extends AbstractIterativeComputation {
-		private final int d_chain;
-		
-		public BurnInChain(int chain) {
-			super(d_burnInIter);
-			d_chain = chain;
-		}
-		
-		public void doStep() {
-			update(d_chain);
-		}
-	}
-
-	private class SimulationChain extends AbstractExtendableIterativeComputation {
-		private final int d_chain;
-		
-		public SimulationChain(int chain) {
-			super(d_simulationIter);
-			d_chain = chain;
-		}
-		
-		public void doStep() {
-			update(d_chain);
-			output(d_chain);
-		}
-	}
-
-	private class BurnInTask extends IterativeTask {
-		public BurnInTask(int chain) {
-			super(new BurnInChain(chain), "Tuning: " + chain);
-			setReportingInterval(d_reportingInterval);
-		}
-	}
-	
-	private class SimulationTask extends ExtendableIterativeTask {
-		public SimulationTask(int chain) {
-			super(new SimulationChain(chain), "Simulation: " + chain);
-			setReportingInterval(d_reportingInterval);
-		}
-	}
-	
-	private class SimpleRestartableSuspendableTask extends SimpleSuspendableTask {
-
-		public SimpleRestartableSuspendableTask(Runnable runnable, String string) {
-			super(runnable, string);
-		} 
-		
-		public void reset() {
-			d_finished = false;
-			d_mgr.fireTaskRestarted();
-		}
-	}
+	private List<Parameter> d_parameters;
 	
 	public YadasModel(Network network) {
+		super();
 		d_network = network;
-
-		buildActivityModel();
 	}
-
-	private void buildActivityModel() {
-		// Create tasks for each phase of the MCMC simulation
-		Task buildModelPhase = new SimpleSuspendableTask(new Runnable() {
-			public void run() {
-				buildModel();
-			}
-		}, STARTING_SIMULATION_PHASE);
-		final List<Task> burnInPhase = new ArrayList<Task>(d_nChains);
-		final List<Task> simulationPhase = new ArrayList<Task>(d_nChains);
-		for (int i = 0; i < d_nChains; ++i) {
-			burnInPhase.add(new BurnInTask(i));
-			simulationPhase.add(new SimulationTask(i));
-		}
-
-		d_extendDecisionPhase = new ExtendDecisionTask();
-				
-		d_extendSimulationPhase = new SimpleRestartableSuspendableTask(new Runnable() {
-			public void run() {
-				// Extend the simulations. This is safe because they won't be started before this task is finished.
-				for(Task t : simulationPhase) {
-					((ExtendableIterativeTask) t).extend(d_simulationIter);
-				}
-				d_results.setNumberOfIterations(d_results.getNumberOfIterations() + d_simulationIter);
-				d_results.setDerivedParameters(getDerivedParameters());
-				// Finally, reset the decision phase. Must be done last otherwise it becomes a next state.
-				((SimpleRestartableSuspendableTask) d_notifyResults).reset();
-				((ExtendDecisionTask) d_extendDecisionPhase).reset();
-			}
-		}, MixedTreatmentComparison.EXTENDING_SIMULATION_PHASE);
-		
-		d_finalPhase = new NullTask();
-		
-		d_notifyResults = new SimpleRestartableSuspendableTask(new Runnable() {	
-			@Override
-			public void run() {
-				d_results.simulationFinished();
-			}
-		}, MixedTreatmentComparison.CALCULATING_SUMMARIES_PHASE);
-		
-		// Build transition graph between phases of the MCMC simulation
-		List<Transition> transitions = new ArrayList<Transition>();
-		transitions.add(new ForkTransition(buildModelPhase, burnInPhase));
-		for (int i = 0; i < d_nChains; ++i) {
-			transitions.add(new DirectTransition(burnInPhase.get(i), simulationPhase.get(i)));
-		}
-		transitions.add(new JoinTransition(simulationPhase, d_notifyResults));
-		transitions.add(new DirectTransition(d_notifyResults, d_extendDecisionPhase));
-		transitions.add(new DecisionTransition(d_extendDecisionPhase, d_extendSimulationPhase, d_finalPhase, new Condition() {
-			public boolean evaluate() {
-				return d_extendSimulation == ExtendSimulation.EXTEND;
-			}
-		}));
-		transitions.add(new ForkTransition(d_extendSimulationPhase, simulationPhase));
-		// Together they form the full "activity"
-		ActivityModel activityModel = new ActivityModel(buildModelPhase, d_finalPhase, transitions);
-		d_activityTask = new ActivityTask(activityModel, "MCMC model");
-	}
-
 
 	abstract protected boolean isInconsistency();
-	
-	public boolean isReady() {
-		return d_activityTask.isFinished() || d_extendDecisionPhase.isStarted();
-	}
-
-	public ActivityTask getActivityTask() {
-		return d_activityTask;
-	}
 	
 	@Override
 	public BasicParameter getRelativeEffect(Treatment base, Treatment subj) {
 		return new BasicParameter(base, subj);
 	}
 
-	public int getBurnInIterations() {
-		return d_burnInIter;
-	}
-
-	public void setBurnInIterations(int it) {
-		validIt(it);
-		d_burnInIter = it;
-		buildActivityModel();
-	}
-
-	public int getSimulationIterations() {
-		return d_simulationIter;
-	}
-
-	public void setSimulationIterations(int it) {
-		validIt(it);
-		d_simulationIter = it;
-		buildActivityModel();
-	}
-
 	@Override
 	public Parameter getRandomEffectsVariance() {
 		return d_randomEffectVar;
-	}
-	
-	public MCMCResults getResults() {
-		return d_results;
-	}
-
-	private void validIt(int it) {
-		if (it <= 0 || it % 100 != 0) {
-			throw new IllegalArgumentException("Specified # iterations should be a positive multiple of 100");
-		}
 	}
 	
 	////
@@ -335,6 +117,29 @@ abstract class YadasModel implements MixedTreatmentComparison {
 	//// Below: code to initialize the MCMC model
 	////
 
+	@Override
+	protected void prepareModel() {
+		d_pmtz = buildNetworkModel();
+		JDKRandomGenerator rng = new JDKRandomGenerator();
+		double scale = VARIANCE_SCALING;
+		for (int i = 0; i < d_nChains; ++i) {
+			d_startGen.add(AbstractDataStartingValueGenerator.create(d_network, NetworkModel.createComparisonGraph(d_network), rng, scale));
+		}
+		
+		d_priorGen = new PriorGenerator(d_network);
+		
+		d_parameters = new ArrayList<Parameter>(d_pmtz.getParameters());
+		d_parameters.add(d_randomEffectVar);
+		if (isInconsistency()) {
+			d_parameters.add(d_inconsistencyVar);
+		}
+	}
+	
+	@Override
+	protected List<Parameter> getParameters() {
+		return d_parameters;
+	}
+	
 	protected abstract Parameterization buildNetworkModel();
 	
 	protected Map<NetworkParameter, Derivation> getDerivedParameters() {
@@ -349,39 +154,13 @@ abstract class YadasModel implements MixedTreatmentComparison {
 		}
 		return map;
 	}
-
-	private void buildModel() {
-		d_pmtz = buildNetworkModel();
-		JDKRandomGenerator rng = new JDKRandomGenerator();
-		double scale = VARIANCE_SCALING;
-		for (int i = 0; i < d_nChains; ++i) {
-			d_startGen.add(AbstractDataStartingValueGenerator.create(d_network, NetworkModel.createComparisonGraph(d_network), rng, scale));
-		}
-		
-		d_priorGen = new PriorGenerator(d_network);
-
-		List<Parameter> parameters = new ArrayList<Parameter>(d_pmtz.getParameters());
-		parameters.add(d_randomEffectVar);
-		if (isInconsistency()) {
-			parameters.add(d_inconsistencyVar);
-		}
-
-		d_results.setDirectParameters(parameters);
-		d_results.setNumberOfChains(d_nChains);
-		d_results.setNumberOfIterations(d_simulationIter);
-
-		d_results.setDerivedParameters(getDerivedParameters());
-
-		for (int i = 0 ; i < d_nChains; ++i) {
-			createChain(i);
-		}
-	}
 	
 	////
 	//// Below: code to create the structure of the MCMC model.
 	////
 
-	private void createChain(int chain) {
+	@Override
+	protected void createChain(int chain) {
 		StartingValueGenerator startVal = d_startGen.get(chain);
 
 		// study baselines
@@ -517,12 +296,7 @@ abstract class YadasModel implements MixedTreatmentComparison {
 				);
 		}
 		
-		List<MCMCUpdate> tuners = new ArrayList<MCMCUpdate>(params.size());
-		for (MCMCParameter param : params) {
-			tuners.add(new UpdateTuner(param, d_burnInIter / 50, 50, 1, Math.exp(-1)));
-		}
-
-		d_updateList.add(tuners);
+		addTuners(params);
 
 		List<ParameterWriter> writers = new ArrayList<ParameterWriter>(params.size());
 		for (int i = 0; i < parameters.size(); ++i) {
@@ -532,8 +306,7 @@ abstract class YadasModel implements MixedTreatmentComparison {
 		if (isInconsistency()) {
 			writers.add(d_results.getParameterWriter(d_inconsistencyVar, chain, sigmaw, 0));
 		}
-
-		d_writeList.add(writers);
+		addWriters(writers);
 	}
 
 	protected int getNumberOfBasicParameters() {
@@ -642,31 +415,6 @@ abstract class YadasModel implements MixedTreatmentComparison {
 
 	private int reDim(Study s) {
 		return s.getTreatments().size() - 1;
-	}
-
-	private void update(int chain) {
-		for (MCMCUpdate u : d_updateList.get(chain)) {
-			try {
-				u.update();
-			} catch(Exception e) {
-				throw new RuntimeException("Failed to update " + u, e);
-			}
-		}
-	}
-
-	protected void output(int chain) {
-		for (ParameterWriter p : d_writeList.get(chain)) {
-			p.output();
-		}
-	}
-
-	public void setExtendSimulation(ExtendSimulation s) {
-		d_extendSimulation = s;
-	}
-
-	public MCMCSettingsCache getSettings() {
-		return new MCMCSettingsCache(d_simulationIter / (2 * THINNING_INTERVAL), d_simulationIter, 
-				THINNING_INTERVAL, d_burnInIter, VARIANCE_SCALING, d_nChains);
 	}
 
 }
